+++ src/tests/demo.test.ts
@@ -1,3 +1,4 @@
+// Add here your tests to see if your app is working as expected
 import { describe, expect, it } from "vitest";
 
 describe("Demo test", () => {