+++ .github/CHANGELOG.md
@@ -21,6 +21,7 @@ This is the initial version of this template, it does have the following:
 - Added a demo test suit, and benchmark test.
 - Added the initial folder structure.
 - Added a healthcheck at `/api` route.
+- Added the common folder structure for the projects, with demo example on each.
 
 ### UI and Components
 - We added the light/dark theme to the template.